fix: resolve compiler warnings across plugin codebase
- ViewTransitionManager.h: remove unused lambda captures, fix int-to-float conversions
- TransitionAnimation.h: fix int64-to-int conversion with explicit cast
- AnimationTimeline.h: fix int64-to-int conversion, add [[maybe_unused]] for params
- OperationalTransform.h: add explicit copy constructor to fix deprecated warning
- Easing.h: fix int-to-char conversion in std::tolower()
- CacheLayer.h: fix incorrect [[maybe_unused]] attribute placement
- PluginEditor.cpp: remove unused 'this' capture from lambda
- SoundPage.cpp: replace deprecated .boldened() with .withStyle("Bold")
- Recording.cpp: fix deprecated Font constructor
- Search.cpp: fix deprecated Font constructor
- AggregatedFeedCard.cpp: rename lambda params to avoid shadowing

+++ plugin/src/ui/animations/TransitionAnimation.h
@@ -253,7 +253,7 @@ public:
     auto now = std::chrono::steady_clock::now();
     auto elapsed =
         std::chrono::duration_cast<std::chrono::milliseconds>(now - startTime_);
-    int elapsedMs = elapsed.count() - pausedElapsed_;
+    int elapsedMs = static_cast<int>(elapsed.count()) - pausedElapsed_;
     return std::max(0, elapsedMs);
   }
